### 2025-11-05 JosephBrunet/cardiotensor#24 @chourroutm
Add OME-Zarr I/O support
With the development of https://github.com/chourroutm/ome-zarr-tools, I would like to push for the use of OME-Zarr across the group as a "standardized" format for I/O.
