### 2026-01-01 #9720 @editorialbot
[REVIEW]: Cardiotensor: A Python Library for Orientation Analysis and Tractography in 3D Cardiac Imaging
**Submitting author:** @JosephBrunet (<a href="http://orcid.org/0000-0002-8424-9510">Joseph Brunet</a>)
**Repository:** https://github.com/JosephBrunet/cardiotensor
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
