### 2026-01-01 #8809 @danielskatz
@finsberg - you shouldn't be a reviewer if you are the editor. You can start the review with just one reviewer, if you want to give that reviewer a chance to get started, but you still need to find a second independent (not you) reviewer in the review thread.
