### 2026-01-01 #9720 @finsberg
@sitic thanks for your help for reviewing this work! As you know, this is where the review happens. I kindly ask you to now formally start the review. Follow the instructions above ☝️ to generate a check box list for yourself here to guide you through the process (i.e adding a comment with the message @editorialbot generate my checklist). Let me know if you have any questions. The reviewers are encouraged to submit issues and pull requests on the software repository. When posting issues on the software repository, I appreciate it if you mention https://github.com/openjournals/joss-reviews/issues/9720 so that a link is created to this thread and I can keep track of what is happening.
