+++ src/cardiotensor/utils/DataReader.py
@@ -1,11 +1,12 @@
 import os
 import sys
-from concurrent.futures import ThreadPoolExecutor, as_completed
 from os import PathLike
 from pathlib import Path
 from typing import Any
 
 import cv2
+from dask import compute, delayed
+import glymur
 import numpy as np
 import psutil
 import SimpleITK as sitk
@@ -49,6 +50,42 @@ def _fit(
             constant_values=pad_value,
         )
     return arr
+
+
+def _read_image_file(file_path: Path) -> np.ndarray:
+    """Reads one image file to a NumPy array."""
+    suffix = file_path.suffix.lower()
+
+    if suffix == ".npy":
+        return np.load(file_path, mmap_mode="r")
+
+    if suffix in {".tif", ".tiff"}:
+        try:
+            return tiff.imread(file_path)
+        except Exception as e:
+            img = cv2.imread(str(file_path), -1)
+            if img is not None:
+                print(
+                    f"⚠ Falling back to OpenCV for TIFF '{file_path}' "
+                    f"(tifffile error: {e})"
+                )
+                return img
+            raise RuntimeError(
+                "Failed to read TIFF with tifffile and OpenCV. "
+                "If the file uses LZW/other codecs, install 'imagecodecs' "
+                "to enable decoding."
+            ) from e
+
+    if suffix == ".jp2":
+        return glymur.Jp2k(str(file_path))[:]
+
+    img = cv2.imread(str(file_path), -1)
+    if img is None:
+        raise RuntimeError(
+            f"cv2.imread failed for '{file_path}'. "
+            "File may be missing or in an unsupported/invalid format."
+        )
+    return img
 
 
 class DataReader:
@@ -295,50 +332,15 @@ class DataReader:
         """
         Reads an image from the given file path into a NumPy array.
         """
-        suffix = file_path.suffix.lower()
-
-        # 1) .npy → memory-mapped (no full copy)
-        if suffix == ".npy":
-            return np.load(file_path, mmap_mode="r")
-
-        # 2) TIFF → prefer tifffile, but gracefully fall back to OpenCV if codecs missing
-        if suffix in {".tif", ".tiff"}:
-            try:
-                # Best path: uses tifffile (and imagecodecs when available)
-                return tiff.imread(file_path)
-            except Exception as e:
-                # Fallback path: try OpenCV (handles many baseline TIFFs, including your tests)
-                img = cv2.imread(str(file_path), -1)
-                if img is not None:
-                    # Optional: warn once so users know performance/features may be reduced
-                    print(
-                        f"⚠ Falling back to OpenCV for TIFF '{file_path}' "
-                        f"(tifffile error: {e})"
-                    )
-                    return img
-                # If even OpenCV fails, raise a helpful error
-                raise RuntimeError(
-                    "Failed to read TIFF with tifffile and OpenCV. "
-                    "If the file uses LZW/other codecs, install 'imagecodecs' "
-                    "to enable decoding."
-                ) from e
-
-        # 3) Other formats → OpenCV (allow any depth/color)
-        img = cv2.imread(str(file_path), -1)
-        if img is None:
-            raise RuntimeError(
-                f"cv2.imread failed for '{file_path}'. "
-                "File may be missing or in an unsupported/invalid format."
-            )
-        return img
+        return _read_image_file(file_path)
 
     def _load_image_stack(
         self, file_list: list[Path], start_index: int, end_index: int
     ) -> np.ndarray:
         """
         Efficiently loads a stack of images into a 3D (or 4D for vector .npy) NumPy array
-        by preallocating the output and filling it in place. Uses a bounded thread pool
-        for I/O bound reads and keeps slice order via indices.
+        by preallocating the output and filling it in place. Uses batched dask
+        multiprocess reads and keeps slice order via indices.
         """
 
[…]
