### 2026-02-19 #9720 @finsberg
Hi @jijoderick @paquiteau @tomelse @jonbmartin @ClementCaporal👋
Considering your expertise, would any of you be willing to review this submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/pre-2026-submission/

If you are not available it would be great if you have any suggestions for potential reviewers.

Paper: https://github.com/openjournals/joss-papers/blob/joss.09720/joss.09720/10.21105.joss.09720.pdf
Software: https://github.com/JosephBrunet/cardiotensor
