### 2026-02-19 #9720 @ClementCaporal
Hello @finsberg, thank you for the proposition.
The paper and software seems interesting and well maintained. I never worked with cardiac imaging, but I love 3D image and angles so i'll do my best to review this submission.

However, due to previous deadlines, I think I might not be able to start before the 10th of Mars...

With a quick look at the paper I don't see the "AI usage disclosure" that seems to be now required?

PS: I like the logo @JosephBrunet !
