### 2026-02-19 #9720 @finsberg
Hi @ClementCaporal, thank you so much for agreeing to review! Starting around March 10th is perfectly fine—I really appreciate you being upfront about your timeline.

Regarding the AI usage disclosure: this submission was received before January 5, 2026, so it falls under our previous guidelines and doesn't require one (see [this blog post](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) for context). I apologize for linking you to the new checklist earlier (it has been updated now)! You can find the correct, archived pre-2026 reviewer documentation here: https://joss.readthedocs.io/en/pre-2026-submission/

I'll go ahead and add you as a reviewer now. Thanks again!
