### 2026-02-19 #9720 @jonbmartin
Yes, I’d be happy to review!

Jonathan Martin

On Thu, Feb 19, 2026 at 2:17 PM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#9720)
> <https://github.com/openjournals/joss-reviews/issues/9720#issuecomment-3929721581>
>
> @ClementCaporal <https://github.com/ClementCaporal> added to the
> reviewers list!
>
