### 2026-02-19 #9720 @finsberg
@ClementCaporal and @jonbmartin thanks for your help for reviewing this work! Feel free to start the review when you are ready.

The review will happen in this issue so it is great if you mention https://github.com/openjournals/joss-reviews/issues/9720 when opening issues in the software repository so that I can keep track of what is happening here.

Please follow the instructions above ☝️ to generate a check box list for yourself here to guide you through the process (i.e adding a comment with the message `@editorialbot generate my checklist`).
