### 2026-02-24 JosephBrunet/cardiotensor#27 @jonbmartin
Documentation typo
Minor typo - I think that in the (very nice!) documentation, in the 'Getting Started' example, under 'Visualizing Transmural Profiles', the command:

`cardio-analysis ./parameters_example.conf 150
`
is suggested. Looks like cardio-analysis won't accept just 150 as an argument on its own, should be proceeded by --N-slice, e.g.

`cardio-analysis ./parameters_example.conf --N-slice 150
`
which is clarified in the usage message.

cardio-analysis doesn't have much documentation on the documentation site - could be helpful to have info on that function beyond just the example
