### 2026-02-24 82a3f1f
fix typo in example
