+++ docs/getting-started/examples.md
@@ -133,9 +133,9 @@ Once the structure tensor and eigenvectors have been calculated on the whole vol
 
 ### Generating and visualizing streamlines
 
-1. Use the `cardio-visualize-vector` command to generate the streamlines:
+1. Use the `cardio-generate-streamlines` command to generate the streamlines:
     ```console
-    $ cardio-generate-streamlines parameters_example.conf --seeds 10000 --start 150
+    $ cardio-generate-streamlines parameters_example.conf --seeds 10000 --start-z 150
     ```
 
     - `--start`
