### 2026-02-25 a5e8428
add docs advanced section
