### 2026-02-26 04b94a9
Add computation of myocardial disarray index (MDI)
