+++ src/cardiotensor/orientation/orientation_computation_pipeline.py
@@ -16,12 +16,14 @@ from cardiotensor.orientation.orientation_computation_functions import (
     compute_azimuth_and_elevation,
     compute_fraction_anisotropy,
     compute_helix_and_transverse_angles,
+    compute_MDI_from_v,
     interpolate_points,
     plot_images,
     remove_padding,
     rotate_vectors_to_new_axis,
     write_images,
     write_vector_field,
+    write_MDI_images
 )
 from cardiotensor.utils.DataReader import DataReader
 from cardiotensor.utils.utils import remove_corrupted_files
@@ -136,6 +138,7 @@ def compute_orientation(
     n_slice_test: int | None = None,
     start_index: int = 0,
     end_index: int | None = None,
+    write_MDI: bool = False
 ) -> None:
     """
     Compute the orientation for a volume dataset.
@@ -298,8 +301,19 @@ Parameters:
 
         del mask
 
+    if write_MDI: 
+        print("\n" + "-" * 40)
+        print("MDI")
+        print("-" * 40 + "\n")
+        print("Calculating myocardial disarray index...")
+        MDI_vol = compute_MDI_from_v(vec, window_size = int(rho*4 + 1))
+        print("MDI computation complete")
+        MDI_vol, _, _ = remove_padding(MDI_vol, val, vec, padding_start, padding_end)
+        print(f"MDI shape after removing padding: {MDI_vol.shape}")
+        
+    
     volume, val, vec = remove_padding(volume, val, vec, padding_start, padding_end)
-    print(f"Vector shape after removing padding: {vec.shape}")
+    print(f"ST Vector shape after removing padding: {vec.shape}")
 
     center_line = center_line[start_index_padded:end_index_padded]
 
@@ -312,7 +326,11 @@ Parameters:
     # vec[:, negative_z] *= -1
 
     t2 = time.perf_counter()  # stop time
-    print(f"finished calculating structure tensors in {t2 - t1} seconds")
+    if write_MDI: 
+        print(f"finished calculating structure tensors and MDI in {t2 - t1} seconds")
+    else:
+        print(f"finished calculating structure tensors in {t2 - t1} seconds")
+    
 
     print("\n" + "-" * 40)
     print("ANGLE & ANISOTROPY CALCULATION")
@@ -353,6 +371,8 @@ Parameters:
                             start_index,
                             write_vectors,
                             write_angles,
+                            write_MDI,
+                            MDI_vol[z, :, :] if write_MDI else None,
                             is_test,
                             angle_mode,
                         ),
@@ -382,6 +402,8 @@ Parameters:
                     start_index,
                     write_vectors,
                     write_angles,
+                    write_MDI,
+                    MDI_vol[z, :, :] if write_MDI else None,
                     is_test,
                     angle_mode,
                 )
@@ -406,6 +428,8 @@ def compute_slice_angles_and_anisotropy(
     start_index: int = 0,
     write_vectors: bool = False,
     write_angles: bool = True,
+    write_MDI: bool = False,
+    MDI_slice: np.ndarray = None,
     is_test: bool = False,
     angle_mode: str = "ha_ia",
 ) -> None:
@@ -439,6 +463,11 @@ def compute_slice_angles_and_anisotropy(
     if write_vectors:
         expected_paths.append(
             os.path.join(output_dir, "eigen_vec", f"eigen_vec_{idx:06d}.npy")
+        )
+    
+    if write_MDI:
+        expected_paths.append(
+            os.path.join(output_dir, "MDI", f"MDI_{idx:06d}.npy")
         )
 
     # Skip if all outputs are already present and we are not in test mode
@@ -475,6 +504,9 @@ def compute_slice_angles_and_anisotropy(
             img_angle1, img_angle2 = compute_azimuth_and_elevation(
[…]
