### 2026-02-27 e972f08
fix: Update cardio-analysis command to use --N-slice flag
