### 2026-02-27 be376f2
simplify and add links
