### 2026-03-02 298ab67
Add MDI parallel Processing
Accelerates computation and allows visual tracking of progress
