+++ src/cardiotensor/orientation/multiprocessing_MDI.py
@@ -1,0 +1,244 @@
+import logging
+import os
+from dataclasses import dataclass
+from multiprocessing import Pool, RawArray, SimpleQueue, cpu_count
+from multiprocessing.pool import ThreadPool
+from typing import Callable, Literal, Sequence
+import numpy as np
+import numpy.typing as npt
+from scipy.ndimage import uniform_filter
+
+from . import util
+
+logger = logging.getLogger(__name__)
+
+DEFAULT_POOL_TYPE = "thread" if os.name == "nt" else "process"
+
+
+# -----------------------------------------------------------------------------
+# Shared memory helpers (unchanged pattern)
+# -----------------------------------------------------------------------------
+
+@dataclass(frozen=True)
+class _ArrayArgs:
+    array: np.ndarray
+    def get_array(self) -> np.ndarray:
+        return self.array
+
+
+@dataclass(frozen=True)
+class _RawArrayArgs:
+    array: any
+    shape: tuple[int, ...]
+    dtype: npt.DTypeLike
+    def get_array(self) -> np.ndarray:
+        return np.frombuffer(self.array, dtype=self.dtype).reshape(self.shape)
+
+
+def _create_raw_array(shape: tuple[int, ...], dtype: npt.DTypeLike):
+    raw = RawArray(
+        "b",
+        np.prod(np.asarray(shape), dtype=np.int64).item()
+        * np.dtype(dtype).itemsize,
+    )
+    a = np.frombuffer(raw, dtype=dtype).reshape(shape)
+    return raw, a
+
+
+# -----------------------------------------------------------------------------
+# Init args
+# -----------------------------------------------------------------------------
+
+@dataclass(frozen=True)
+class _InitArgs:
+    data_args: _RawArrayArgs | _ArrayArgs
+    output_args: _RawArrayArgs | _ArrayArgs
+    window_size: int
+    block_size: int
+    devices: SimpleQueue
+
+    def get_data_sources(self):
+        return (
+            self.data_args.get_array(),
+            self.output_args.get_array(),
+            self.devices.get(),
+        )
+
+
+_worker_args = None
+_data = None
+_output = None
+_device = None
+
+
+# -----------------------------------------------------------------------------
+# Worker init
+# -----------------------------------------------------------------------------
+
+def _init_worker(init_args: _InitArgs):
+    global _worker_args, _data, _output, _device
+    _worker_args = init_args
+    _data, _output, _device = init_args.get_data_sources()
+
+
+# -----------------------------------------------------------------------------
+# Core MDI computation (block-level)
+# -----------------------------------------------------------------------------
+
+def _compute_mdi_block(block: np.ndarray, window_size: int):
+
+    vx, vy, vz = block
+
+    norm = np.sqrt(vx**2 + vy**2 + vz**2)
+    mask = norm > 0
+
+    vx = np.where(mask, vx / norm, 0)
+    vy = np.where(mask, vy / norm, 0)
+    vz = np.where(mask, vz / norm, 0)
+
+    Sxx = uniform_filter(vx * vx, size=window_size)
+    Syy = uniform_filter(vy * vy, size=window_size)
+    Szz = uniform_filter(vz * vz, size=window_size)
+
+    Sxy = uniform_filter(vx * vy, size=window_size)
+    Sxz = uniform_filter(vx * vz, size=window_size)
+    Syz = uniform_filter(vy * vz, size=window_size)
+
+    S = np.stack([
+        np.stack([Sxx, Sxy, Sxz], axis=-1),
+        np.stack([Sxy, Syy, Syz], axis=-1),
+        np.stack([Sxz, Syz, Szz], axis=-1),
+    ], axis=-2)
+
[…]
