### 2026-03-03 d6bda01
fix default colormap
