### 2026-03-03 317dd38
add colormap arg
