+++ src/cardiotensor/orientation/orientation_computation_functions.py
@@ -502,140 +502,6 @@ def plot_images(
     plt.show()
 
 
-# def write_images(
-#     img_angle1: np.ndarray,
-#     img_angle2: np.ndarray,
-#     img_FA: np.ndarray,
-#     start_index: int,
-#     output_dir: str,
-#     output_format: str,
-#     output_type: str,
-#     z: int,
-#     colormap_angle=None,
-#     colormap_FA=None,
-#     angle_names: tuple[str, str] = ("HA", "IA"),
-#     angle_ranges: tuple[tuple[float, float], tuple[float, float]] = ((-90, 90), (-90, 90)),
-# ) -> None:
-
-#     # angle_names drives subfolders, angle_ranges drives normalization
-#     name1, name2 = angle_names
-#     (a1_min, a1_max), (a2_min, a2_max) = angle_ranges
-
-#     os.makedirs(f"{output_dir}/{name1}", exist_ok=True)
-#     os.makedirs(f"{output_dir}/{name2}", exist_ok=True)
-#     os.makedirs(f"{output_dir}/FA", exist_ok=True)
-
-#     if "8bit" in output_type:
-#         img1_8 = convert_to_8bit(img_angle1, min_value=a1_min, max_value=a1_max)
-#         img2_8 = convert_to_8bit(img_angle2, min_value=a2_min, max_value=a2_max)
-#         imgFA_8 = convert_to_8bit(img_FA, min_value=0, max_value=1)
-
-#         if output_format == "jp2":
-#             # same glymur writes, just using name1/name2
-#             glymur.Jp2k(f"{output_dir}/{name1}/{name1}_{(start_index+z):06d}.jp2",
-#                         data=img1_8, cratios=[10], numres=8, irreversible=True)
-#             glymur.Jp2k(f"{output_dir}/{name2}/{name2}_{(start_index+z):06d}.jp2",
-#                         data=img2_8, cratios=[10], numres=8, irreversible=True)
-#             glymur.Jp2k(f"{output_dir}/FA/FA_{(start_index+z):06d}.jp2",
-#                         data=imgFA_8, cratios=[10], numres=8, irreversible=True)
-#         elif output_format == "tif":
-#             tifffile.imwrite(f"{output_dir}/{name1}/{name1}_{(start_index+z):06d}.tif", img1_8)
-#             tifffile.imwrite(f"{output_dir}/{name2}/{name2}_{(start_index+z):06d}.tif", img2_8)
-#             tifffile.imwrite(f"{output_dir}/FA/FA_{(start_index+z):06d}.tif", imgFA_8)
-#         else:
-#             sys.exit(f"I don't recognise the output_format ({output_format})")
-
-
-#     # ---- RGB output ----
-#     elif "rgb" in output_type:
-
-#         def write_img_rgb(
-#             img: np.ndarray,
-#             output_path: str,
-#             cmap: plt.Colormap,
-#             vmin: float,
-#             vmax: float,
-#         ) -> None:
-#             """
-#             Writes a single 2D RGB image using a fixed colormap range.
-
-#             Args:
-#                 img (np.ndarray): Input scalar image.
-#                 output_path (str): Path to save the RGB image.
-#                 cmap (plt.Colormap): Matplotlib colormap (e.g., inferno, custom HA cmap).
-#                 vmin (float): Minimum value for normalization.
-#                 vmax (float): Maximum value for normalization.
-#             """
-#             img_clipped = np.clip(img, vmin, vmax)
-#             img_norm = (img_clipped - vmin) / (vmax - vmin + 1e-8)
-
-#             img_rgb = cmap(img_norm)[..., :3]  # Drop alpha channel
[…]
