### 2026-03-03 #9720 @ClementCaporal
Hello,
I started to review the paper and repo today.
It was very easy to dive in the code and check the example, so thanks to @JosephBrunet , et al for the time they took to polish their package and documentation.

- In the paper there are some claims about "very large 3D" datasets, so I will take some time to check the `Functionality/Performance` box. I will use https://human-organ-atlas.esrf.fr/datasets/1659197537 but I'll do it next week.
_Related suggestion_: I don't know how "very large" will be interpreted through time, in your paper maybe you could explicit either the dataset size you tested ("Tb large dataset") or replace with "arbitrary large dataset". If you think it is relevant, an estimation of the time per gpu for 1 TB dataset could be useful for a first glance estimation for new users. (Note that this is _only a suggestion_ and I think your contribution meets all the criteria for publication)
In the same memory theme, I look forward to how you will integrate ome-zarr I/O support to your tool (https://github.com/JosephBrunet/cardiotensor/issues/24).

- Launcher Documentation: I see you have scripts and function to use cardiotensor with slurm. I think this feature should highlighted somewhere in your docs or readme. (And it would even help me to check quickly the large 3D dataset claim!)
