### 2026-03-07 #9720 @JosephBrunet
Dear @ClementCaporal,

Thank you for taking the time to review this

You are right that the term “very large 3D datasets” was a bit vague. I have clarified this in the paper by specifying that cardiotensor has been applied to datasets of about 1–5 TB (teravoxel scale).

For the SLURM launcher, I have now improved the launcher script and added a doc page explaining how to run cardiotensor with SLURM.

Concerning ome-zarr, I agree it would be very useful, especially for memory management on large datasets, it is on the roadmap!

Thanks again for the helpful suggestions
