### 2026-03-23 #9720 @JosephBrunet
@ClementCaporal, @jonbmartin,

Were you able to make progress on the review? Please let me know if you need anything from my side.

Looking forward to your feedback.

Best,
Joseph
