### 2026-03-27 #9720 @finsberg
Hi @ClementCaporal and @jonbmartin, thanks again for taking the time to review this. We really appreciate your valuable feedback!

@ClementCaporal, I see the checkbox for "performance" hasn't been ticked off yet. Could you let me know what still needs to be addressed there?

@jonbmartin, just checking in on the last few checkboxes for your review. Do you have a rough idea of when you'll be able to wrap things up?

Thanks again for your contributions!
