### 2026-03-31 1ea1255
fix mask path bug
