### 2026-04-01 d4f4920
simplify call function for testing
