### 2026-04-01 b762998
add colormap to quiver
