### 2026-04-03 #9720 @jonbmartin
Hi @finsberg and @ClementCaporal, I've finished my review. This is an excellent toolbox and I believe that the minimum requirements for publication listed above are all satisfied.

Cardiotensor is a Python package for quantifying 3D cardiomyocite orientation in large volumetric cardiac datasets. Unlike existing tools (MRtrix3, DIPY), it operates on direct image intensity gradients rather than diffusion modelling, making it applicable to synchrotron tomography, micro-CT, and optical microscopy. It computes helix angle (HA), intrusion/transverse angle, and fractional anisotropy (FA), and supports streamline tractography. Its chunk-based parallel architecture supports teravoxel-scale datasets with SLURM integration. This is a great contribution to the field, and the prior validation paper (Brunet et. al. Radiology 2024) is very encouraging.

**DISCLOSURE: CLAUDE AI ASSISTED WITH SOME ASPECTS OF ANALYZING  THE SOURCE CODE** Review was conducted by: reading all source code, reading `paper.md`, reviewing CI configuration, **installing and running the full pipeline on a Raspberry Pi 5 (ARM64, Python 3.13, v1.1.6 codebase) and on a Linux server with GPU(X86_64, Python3.8.10, v1.1.7 codebase)**, executing the provided example dataset end-to-end, and measuring test coverage.

I have a couple of minor critiques that **I do not consider to be blocks for publication**, but should ideally be addressed. I'll put these together as issues.

- Low test coverage: only 13% currently. Places like generate_streamlines.py, slurm_launcher.py have 0% test coverage. Should be expanded
- Silent data loss: In orientation_computation_pipeline.py, the function compute_slice_angles_and_anisotropy() handles test mode at lines 521-555. The test mode branch calls plot_images() and write_images(), then returns. The write_vector_field() call at lines 569-570 is only reached in the non-test branch. No warning is emitted when WRITE_VECTORS=True but is_test=True. Should be fixed or warned about.
- tests.yml only runs on ubuntu-latest with Python 3.11. No macOS, no Windows, no other versions of Python than 3.11
- No public API surface. src/cardiotensor/init.py is empty. If a user imports cardiotensor as a library rather than CLI, they wouldn't have a guided import path so things like 'from cardiotensor import compute_orientation' won't work
- remove_corrupted_files() - deletes anything smaller than 200 bytes. Feels like it could delete small but legitimate files.

My minor issue https://github.com/JosephBrunet/cardiotensor/issues/27 was resolved and fixed already!

Again, recommending publication!
