+++ tests/test_vector_vtk_export.py
@@ -1,0 +1,140 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+import numpy as np
+import pytest
+
+from cardiotensor.utils.vector_vtk_export import (
+    _writeFieldInVtk,
+    export_vector_field_to_vtk,
+    writeStructuredVTK,
+)
+
+
+# ---------------------------------------------------------------------------
+# writeStructuredVTK
+# ---------------------------------------------------------------------------
+
+def test_writeStructuredVTK_scalar(tmp_path: Path):
+    out = str(tmp_path / "out.vtk")
+    data = np.ones((2, 3, 4), dtype=np.float32)
+    writeStructuredVTK(cellData={"myScalar": data}, fileName=out)
+    content = Path(out).read_text()
+    assert "vtk DataFile" in content
+    assert "SCALARS myScalar" in content
+    assert "CELL_DATA" in content
+
+
+def test_writeStructuredVTK_vector(tmp_path: Path):
+    out = str(tmp_path / "out.vtk")
+    data = np.ones((2, 3, 4, 3), dtype=np.float32)
+    writeStructuredVTK(cellData={"vecs": data}, fileName=out)
+    content = Path(out).read_text()
+    assert "VECTORS vecs" in content
+
+
+def test_writeStructuredVTK_no_data(tmp_path: Path, capsys):
+    out = str(tmp_path / "out.vtk")
+    writeStructuredVTK(fileName=out)
+    captured = capsys.readouterr()
+    assert "No data" in captured.out
+
+
+def test_writeStructuredVTK_inconsistent_shapes(tmp_path: Path):
+    out = str(tmp_path / "out.vtk")
+    a = np.ones((2, 3, 4), dtype=np.float32)
+    b = np.ones((2, 3, 5), dtype=np.float32)
+    with pytest.raises(ValueError, match="Inconsistent shape"):
+        writeStructuredVTK(cellData={"a": a, "b": b}, fileName=out)
+
+
+def test_writeStructuredVTK_dimensions(tmp_path: Path):
+    out = str(tmp_path / "out.vtk")
+    data = np.zeros((2, 3, 4), dtype=np.float32)
+    writeStructuredVTK(cellData={"x": data}, fileName=out)
+    content = Path(out).read_text()
+    # dimensions = shape + 1 in each axis
+    assert "DIMENSIONS" in content
+
+
+# ---------------------------------------------------------------------------
+# _writeFieldInVtk
+# ---------------------------------------------------------------------------
+
+def test_writeFieldInVtk_scalar(tmp_path: Path):
+    out = tmp_path / "field.txt"
+    with open(out, "w") as f:
+        _writeFieldInVtk({"s": np.ones((2, 2, 2))}, f)
+    content = out.read_text()
+    assert "SCALARS s float" in content
+    assert "LOOKUP_TABLE default" in content
+
+
+def test_writeFieldInVtk_vector(tmp_path: Path):
+    out = tmp_path / "field.txt"
+    data = np.ones((2, 2, 2, 3))
+    with open(out, "w") as f:
+        _writeFieldInVtk({"v": data}, f)
+    content = out.read_text()
+    assert "VECTORS v float" in content
+
+
+def test_writeFieldInVtk_unsupported_shape(tmp_path: Path):
+    out = tmp_path / "field.txt"
+    with open(out, "w") as f:
+        with pytest.raises(ValueError, match="Unsupported shape"):
+            _writeFieldInVtk({"bad": np.ones((2, 3))}, f)
+
+
+def test_writeFieldInVtk_flat_raises(tmp_path: Path):
+    out = tmp_path / "field.txt"
+    with open(out, "w") as f:
[…]
