+++ tests/test_streamlines_io_utils.py
@@ -1,0 +1,187 @@
+from __future__ import annotations
+
+import numpy as np
+import pytest
+
+from cardiotensor.utils.streamlines_io_utils import (
+    compute_elevation_angles,
+    ha_to_degrees_per_streamline,
+    load_npz_streamlines,
+    normalize_attrs_to_degrees,
+    reduce_per_edge,
+)
+
+
+# ---------------------------------------------------------------------------
+# load_npz_streamlines
+# ---------------------------------------------------------------------------
+
+def _make_npz(tmp_path, streamlines_zyx, extras=None):
+    """Save a .npz with streamlines stored as object array (z, y, x)."""
+    obj = np.empty(len(streamlines_zyx), dtype=object)
+    for i, sl in enumerate(streamlines_zyx):
+        obj[i] = np.asarray(sl, dtype=np.float32)
+    data = {"streamlines": obj}
+    if extras:
+        data.update(extras)
+    p = tmp_path / "test.npz"
+    np.savez(p, **data)
+    return p
+
+
+def test_load_npz_basic(tmp_path):
+    sls_zyx = [
+        [[0, 1, 2], [3, 4, 5]],
+        [[6, 7, 8], [9, 10, 11]],
+    ]
+    p = _make_npz(tmp_path, sls_zyx)
+    streamlines, per_point = load_npz_streamlines(p)
+    assert len(streamlines) == 2
+    assert streamlines[0].shape == (2, 3)
+    # z,y,x → x,y,z swap: first point was (z=0,y=1,x=2) → (x=2,y=1,z=0)
+    np.testing.assert_allclose(streamlines[0][0], [2, 1, 0])
+    assert per_point == {}
+
+
+def test_load_npz_with_values(tmp_path):
+    sls_zyx = [[[0, 0, 0], [1, 1, 1]], [[2, 2, 2], [3, 3, 3]]]
+    ha_obj = np.empty(2, dtype=object)
+    ha_obj[0] = np.array([100.0, 150.0])
+    ha_obj[1] = np.array([50.0, 200.0])
+    p = _make_npz(tmp_path, sls_zyx, extras={"ha_values": ha_obj})
+    streamlines, per_point = load_npz_streamlines(p)
+    assert "HA" in per_point
+    assert len(per_point["HA"]) == 2
+
+
+def test_load_npz_missing_key(tmp_path):
+    p = tmp_path / "empty.npz"
+    np.savez(p, other=np.array([1, 2]))
+    with pytest.raises(ValueError, match="streamlines"):
+        load_npz_streamlines(p)
+
+
+def test_load_npz_values_length_mismatch(tmp_path):
+    sls_zyx = [[[0, 0, 0], [1, 1, 1]]]
+    ha_obj = np.empty(2, dtype=object)  # 2 entries but only 1 streamline
+    ha_obj[0] = np.array([1.0])
+    ha_obj[1] = np.array([2.0])
+    p = _make_npz(tmp_path, sls_zyx, extras={"ha_values": ha_obj})
+    with pytest.raises(ValueError, match="does not match"):
+        load_npz_streamlines(p)
+
+
+# ---------------------------------------------------------------------------
+# ha_to_degrees_per_streamline
+# ---------------------------------------------------------------------------
+
+def test_ha_to_degrees_already_degrees():
+    ha = [np.array([-90.0, 0.0, 90.0], dtype=np.float32)]
+    result = ha_to_degrees_per_streamline(ha)
+    # max is 90 > 1.5 → treated as byte-scaled!
+    # the function checks nanmax > 1.5 so -90..90 will be treated as scaled
+    # just ensure output is same shape
+    assert result[0].shape == ha[0].shape
+
+
+def test_ha_to_degrees_byte_scaled():
+    # 0..255 range → should convert to -90..90
[…]

[…]
