+++ tests/test_downsampling.py
@@ -4,17 +4,16 @@ import math
 from pathlib import Path
 
 import numpy as np
-import pytest
 
 from cardiotensor.utils.downsampling import (
     _process_chunk,
     process_vector_block,
 )
 
-
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
+
 
 def _write_npy_slices(directory: Path, n: int, shape=(3, 8, 8)) -> list[Path]:
     """Write n random (3, H, W) .npy slices and return sorted paths."""
@@ -35,12 +34,15 @@ def _write_npy_slices(directory: Path, n: int, shape=(3, 8, 8)) -> list[Path]:
 # process_vector_block
 # ---------------------------------------------------------------------------
 
+
 def test_process_vector_block_creates_file(tmp_path: Path):
     H, W = 8, 8
     bin_factor = 2
     paths = _write_npy_slices(tmp_path / "input", n=bin_factor)
     out_dir = tmp_path / "output"
-    process_vector_block(paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=0)
+    process_vector_block(
+        paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=0
+    )
     out_file = out_dir / "eigen_vec" / "eigen_vec_000000.npy"
     assert out_file.exists()
     result = np.load(out_file)
@@ -57,7 +59,9 @@ def test_process_vector_block_skips_existing(tmp_path: Path):
     # Write a sentinel file
     sentinel = np.zeros((3, H // bin_factor, W // bin_factor), dtype=np.float32)
     np.save(out_file, sentinel)
-    process_vector_block(paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=0)
+    process_vector_block(
+        paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=0
+    )
     # File should still be the sentinel (not overwritten)
     result = np.load(out_file)
     np.testing.assert_array_equal(result, sentinel)
@@ -68,7 +72,9 @@ def test_process_vector_block_output_dtype(tmp_path: Path):
     bin_factor = 2
     paths = _write_npy_slices(tmp_path / "input", n=bin_factor, shape=(3, H, W))
     out_dir = tmp_path / "output"
-    process_vector_block(paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=3)
+    process_vector_block(
+        paths, bin_factor=bin_factor, h=H, w=W, output_dir=out_dir, idx=3
+    )
     out_file = out_dir / "eigen_vec" / "eigen_vec_000003.npy"
     result = np.load(out_file)
     assert result.dtype == np.float32
@@ -78,13 +84,16 @@ def test_process_vector_block_output_dtype(tmp_path: Path):
 # _process_chunk
 # ---------------------------------------------------------------------------
 
+
 def test_process_chunk_creates_file(tmp_path: Path):
     H, W = 8, 8
     bin_factor = 2
     paths = _write_npy_slices(tmp_path / "input", n=bin_factor, shape=(3, H, W))
     out_dir = tmp_path / "output"
     out_dir.mkdir()
-    _process_chunk(paths, bin_factor=bin_factor, H=H, W=W, eig_out_dir=out_dir, block_idx=0)
+    _process_chunk(
+        paths, bin_factor=bin_factor, H=H, W=W, eig_out_dir=out_dir, block_idx=0
+    )
[…]

[…]
