### 2026-04-06 929f1eb
remove codacy due to weird rules in tests
