### 2026-04-06 c2487ce
add CITATION.cff
