+++ src/cardiotensor/orientation/__init__.py
@@ -1,0 +1,19 @@
+"""Public orientation API."""
+
+from cardiotensor.orientation.orientation_computation_functions import (
+    calculate_structure_tensor,
+    compute_azimuth_and_elevation,
+    compute_fraction_anisotropy,
+    compute_helix_and_transverse_angles,
+    rotate_vectors_to_new_axis,
+)
+from cardiotensor.orientation.orientation_computation_pipeline import compute_orientation
+
+__all__ = [
+    "calculate_structure_tensor",
+    "compute_azimuth_and_elevation",
+    "compute_fraction_anisotropy",
+    "compute_helix_and_transverse_angles",
+    "compute_orientation",
+    "rotate_vectors_to_new_axis",
+]
