+++ src/cardiotensor/analysis/__init__.py
@@ -1,0 +1,15 @@
+"""Public analysis API."""
+
+from cardiotensor.analysis.analysis_functions import (
+    calculate_intensities,
+    find_end_points,
+    plot_intensity,
+    save_intensity,
+)
+
+__all__ = [
+    "calculate_intensities",
+    "find_end_points",
+    "plot_intensity",
+    "save_intensity",
+]
