+++ src/cardiotensor/utils/__init__.py
@@ -1,0 +1,20 @@
+"""Public utility API."""
+
+from cardiotensor.utils.am_utils import write_spatialgraph_am
+from cardiotensor.utils.DataReader import DataReader
+from cardiotensor.utils.streamlines_io_utils import (
+    load_npz_streamlines,
+    load_trk_streamlines,
+)
+from cardiotensor.utils.utils import convert_to_8bit, read_conf_file
+from cardiotensor.utils.vector_vtk_export import export_vector_field_to_vtk
+
+__all__ = [
+    "convert_to_8bit",
+    "DataReader",
+    "export_vector_field_to_vtk",
+    "load_npz_streamlines",
+    "load_trk_streamlines",
+    "read_conf_file",
+    "write_spatialgraph_am",
+]
