+++ src/cardiotensor/__init__.py
@@ -1,11 +1,11 @@
 """
 cardiotensor — 3D cardiomyocyte orientation analysis from volumetric imaging.
 
-Core usage
-----------
->>> from cardiotensor import compute_orientation, DataReader
+Recommended imports
+-------------------
+>>> from cardiotensor import DataReader, compute_orientation, read_conf_file
 >>> reader = DataReader("path/to/images")
->>> compute_orientation("path/to/images", output_dir="output", sigma=3.0, rho=1.0)
+>>> compute_orientation("path/to/images", output_dir="output", sigma=1.0, rho=3.0)
 
 Sub-module imports
 ------------------
@@ -35,37 +35,31 @@ For visualization (requires a display)::
     from cardiotensor.visualization import visualize_streamlines, visualize_vector_field
 """
 
-# --- Core ---
-from cardiotensor.orientation.orientation_computation_pipeline import compute_orientation
-from cardiotensor.utils.DataReader import DataReader
-from cardiotensor.utils.utils import convert_to_8bit, read_conf_file
-
-# --- Orientation ---
-from cardiotensor.orientation.orientation_computation_functions import (
+from cardiotensor.analysis import (
+    calculate_intensities,
+    find_end_points,
+    plot_intensity,
+    save_intensity,
+)
+from cardiotensor.orientation import (
     calculate_structure_tensor,
+    compute_azimuth_and_elevation,
     compute_fraction_anisotropy,
     compute_helix_and_transverse_angles,
+    compute_orientation,
 )
-
-# --- Tractography ---
-from cardiotensor.tractography.generate_streamlines import (
+from cardiotensor.tractography import (
     generate_streamlines_from_params,
     generate_streamlines_from_vector_field,
 )
-
-# --- I/O ---
-from cardiotensor.utils.am_utils import write_spatialgraph_am
-from cardiotensor.utils.streamlines_io_utils import (
+from cardiotensor.utils import (
+    DataReader,
+    convert_to_8bit,
+    export_vector_field_to_vtk,
     load_npz_streamlines,
     load_trk_streamlines,
-)
-from cardiotensor.utils.vector_vtk_export import export_vector_field_to_vtk
-
-# --- Analysis ---
-from cardiotensor.analysis.analysis_functions import (
-    calculate_intensities,
-    find_end_points,
-    save_intensity,
+    read_conf_file,
+    write_spatialgraph_am,
 )
 
 __all__ = [
@@ -76,6 +70,7 @@ __all__ = [
     "convert_to_8bit",
     # Orientation
     "calculate_structure_tensor",
+    "compute_azimuth_and_elevation",
     "compute_fraction_anisotropy",
     "compute_helix_and_transverse_angles",
     # Tractography
@@ -89,5 +84,6 @@ __all__ = [
     # Analysis
     "calculate_intensities",
     "find_end_points",
+    "plot_intensity",
     "save_intensity",
 ]
