+++ src/cardiotensor/tractography/__init__.py
@@ -1,0 +1,11 @@
+"""Public tractography API."""
+
+from cardiotensor.tractography.generate_streamlines import (
+    generate_streamlines_from_params,
+    generate_streamlines_from_vector_field,
+)
+
+__all__ = [
+    "generate_streamlines_from_params",
+    "generate_streamlines_from_vector_field",
+]
