+++ pyproject.toml
@@ -46,7 +46,7 @@ license = {file = "LICENSE"}
 name = "cardiotensor"
 readme = "README.md"
 requires-python = ">=3.10"
-version = "1.1.7"
+version = "1.1.8"
 
 [project.optional-dependencies]
 cupy = ["cupy>=8"]
