+++ mkdocs.yml
@@ -32,6 +32,7 @@ nav:
     - Tractography: getting-technical/tractography.md
   - Advanced:
     - advanced/index.md
+    - Python API: advanced/python-api.md
     - Transmural Analysis: advanced/cardio-analysis.md
     - Streamlines: advanced/streamlines.md
     - SLURM Launcher: advanced/slurm-launcher.md
