+++ docs/advanced/index.md
@@ -1,9 +1,13 @@
 # Advanced
 
-This section is for users who already run the standard `cardio-tensor` pipeline and want more detailed information of the analysis tools.
+This section is for users who already run the standard `cardio-tensor` pipeline
+and want more detailed information about the analysis tools, Python usage, and
+advanced execution workflows.
 
 Pages:
 
+- [Python API](./python-api.md): Use `cardiotensor` as a Python library for
+  structure tensor analysis, angle computation, tractography, and plotting.
 - [Transmural Analysis](./cardio-analysis.md): Extract transmural profiles from angle and FA maps with `cardio-analysis`.
 - [Streamlines](./streamlines.md): Generate and visualize 3D streamlines with `cardio-generate-streamlines` and `cardio-visualize-streamlines`.
 - [SLURM Launcher](./slurm-launcher.md): Submit chunked `cardio-tensor` jobs on HPC clusters.
