### 2026-04-06 #9720 @JosephBrunet
Hi all,

@jonbmartin, thanks a lot for the thorough review and the positive feedback, I’ve gone through the comments and I fully agree with the points raised. I’ll be addressing them in the upcoming updates.

Regarding the post-review checklist:

The new release is v1.1.8, and the archived version is available on Zenodo with DOI: https://doi.org/10.5281/zenodo.19444615

Please let me know if anything else is needed.

Best,
Joseph
