### 2026-04-07 #9720 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
