### 2026-04-28 b0244df
fix if mask not defined
